requirements.txt
setuptools
build
virtualenv
git+https://git.ars-virtualis.org/yul/utils@master
pyyaml
